requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.0
hydra-core==1.3.2
idna==3.6
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
keyboard==0.13.5
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.196
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
